M005: close verification gaps — clippy fixes, MSRV compat, package inventory
Fix all pre-existing clippy errors blocking scripts/verify.sh full:

- codegg-core: replace is_none_or (1.82+) with map_or for MSRV 1.81 compat
- codegg lib: fix ~41 clippy warnings (unused imports, needless borrows,
  redundant closures, derivable impls, field-reassign-with-default,
  suspicious-open-options, large-enum-variant, result-large-err, etc.)
- test files: fix dead-code warnings, constant assertions, unused variables
- Add formal package inventory with crates.io name availability evidence

Evidence:
- scripts/verify.sh quick: EXIT_CODE=0
- scripts/verify.sh full: clippy (-D warnings) passes, 4100 tests pass,
  feature check (server,plugins,lsp-test-support) passes
- 10 crate names available on crates.io (404 confirmed)
- Leaf crates pass cargo package and cargo publish --dry-run
- 1 pre-existing test failure (agent_loop_harness) outside M005 scope

+++ tests/session_projection_m4_controller.rs
@@ -71,10 +71,10 @@ impl ProjectionDigest {
             active_turn_status: active_turn.map(|t| t.status),
             message_count: active_turn.map(|t| t.messages.len()).unwrap_or(0),
             tool_count: active_turn.map(|t| t.tools.len()).unwrap_or(0),
-            pending_permissions: snapshot.primary_session.pending_permission_count as usize,
-            pending_questions: snapshot.primary_session.pending_question_count as usize,
+            pending_permissions: snapshot.primary_session.pending_permission_count,
+            pending_questions: snapshot.primary_session.pending_question_count,
             recent_turn_count: snapshot.recent_turns.len(),
-            active_subagents: snapshot.primary_session.active_subagents as usize,
+            active_subagents: snapshot.primary_session.active_subagents,
             agent_tree: active_turn
                 .map(|t| t.agent_tree.iter().map(|n| (n.task_id, n.status)).collect())
                 .unwrap_or_default(),
